requirements.txt
absl-py==1.3.0
adabound==0.0.5
addict==2.4.0
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.1.7
comm==0.1.2
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distinctipy==1.2.3
docker-pycreds==0.4.0
dotmap==1.3.30
einops==0.5.0
executing==1.2.0
ffmpeg==1.4
filelock==3.8.0
fonttools==4.37.4
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.13.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.50.0
h5py==3.10.0
huggingface-hub==0.10.1
idna==3.4
imageio==2.32.0
importlib-metadata==5.0.0
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.3.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
lmdb==1.4.1
loguru==0.7.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mmcv-full @ file:///home/lrh/mmcv_full-1.5.2-cp38-cp38-manylinux1_x86_64.whl
msgpack==1.0.7
msgpack-numpy==0.4.8
multimethod==1.10
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==25.0.0
randaugment==1.0.2
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.9.2
seaborn==0.12.1
sentry-sdk==1.38.0
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorpack==0.11
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.6.11
toml==0.10.2
torch==1.10.1+cu111
torchinfo==1.7.2
torchpack==0.2.2
torchvision==0.11.2+cu111
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.16.0
wcwidth==0.2.12
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
zipp==3.9.0
